% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_plot.R
\name{discrete_plot}
\alias{discrete_plot}
\title{discrete_plot}
\usage{
discrete_plot(
  data,
  vector,
  plot,
  color = 1,
  title = "discrete plot",
  xlabel = NULL,
  ylabel = NULL,
  xlim = NULL,
  ylim = NULL
)
}
\description{
This function visualizes and summarizes discrete data.
}
\examples{

discrete_plot(iris, iris$Sepal.Length, "bar", "blue", xlim=c(5,7))
}
