% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoplot_same.R
\name{twoplot_same}
\alias{twoplot_same}
\title{twoplot_same}
\usage{
twoplot_same(data, vector1, plot1, vector2, plot2)
}
\description{
This function makes two plots at the same space and summarizes each data. Two vectors are same data
}
\examples{

twoplot_same(iris, iris$Sepal.Length, "point", iris$Petal.Length, "point")
}
