% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoplot_sepa.R
\name{twoplot_sepa}
\alias{twoplot_sepa}
\title{twoplot_sepa}
\usage{
twoplot_sepa(data1, x1, plot1, data2, x2, plot2)
}
\description{
This function makes two plots and shows separated two plots at the same time and summarizes each data.
}
\examples{

twoplot_sepa(iris,iris$Sepal.Length, "point", iris,iris$Petal.Length, "box")
}
